backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# ── Multi-Agent (Microsoft AutoGen v0.2 API) ───────────────────────────────────
ag2>=0.7.5

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0
alembic>=1.13.0

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ── File Upload & Resume Parsing ──────────────────────────────────────────────
python-multipart>=0.0.9
pdfplumber>=0.10.0

# ── Config & Utilities ────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic[email]>=2.5.0
email-validator>=2.1.0
httpx>=0.26.0

# ── Logging ───────────────────────────────────────────────────────────────────
loguru>=0.7.0

# ── Rate Limiting ─────────────────────────────────────────────────────────────
slowapi>=0.1.9

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
